huggingface_space/requirements.txt
torch==2.1.0
torchvision==0.16.0
gradio==4.44.0
Pillow==10.1.0
numpy==1.24.3